feat: Add `preco_unitario` and `data_atualizacao` columns to the `material` table.

+++ migrations/versions/c0f0b9fa83c0_add_preco_unitario_and_data_atualizacao_to_material.py
@@ -20,12 +20,17 @@ def upgrade():
     inspector = sa.inspect(bind)
     with op.batch_alter_table('material', schema=None) as batch_op:
         existing_cols = {col['name'] for col in inspector.get_columns("material")}
-        existing_fks = {fk['name'] for fk in inspector.get_foreign_keys("material")}
         if "preco_unitario" not in existing_cols:
             batch_op.add_column(sa.Column('preco_unitario', sa.Float(), nullable=True))
-        batch_op.add_column(
-            sa.Column('data_atualizacao', sa.DateTime(), server_default=sa.func.now(), nullable=True)
-        )
+        if "data_atualizacao" not in existing_cols:
+            batch_op.add_column(
+                sa.Column(
+                    'data_atualizacao',
+                    sa.DateTime(),
+                    server_default=sa.func.now(),
+                    nullable=True,
+                )
+            )
 
 
 def downgrade():
@@ -33,7 +38,7 @@ def downgrade():
     inspector = sa.inspect(bind)
     with op.batch_alter_table('material', schema=None) as batch_op:
         existing_cols = {col['name'] for col in inspector.get_columns("material")}
-        existing_fks = {fk['name'] for fk in inspector.get_foreign_keys("material")}
-        batch_op.drop_column('data_atualizacao')
-        batch_op.drop_column('preco_unitario')
-
+        if "data_atualizacao" in existing_cols:
+            batch_op.drop_column('data_atualizacao')
+        if "preco_unitario" in existing_cols:
+            batch_op.drop_column('preco_unitario')
